fix sql and db connection from controller
