+++ app/.air.toml
@@ -48,4 +48,4 @@ runner = "green"
 
 [misc]
 # Delete tmp directory on exit
-clean_on_exit = false
+clean_on_exit = true
