+++ app/src/application/repositories/db_repository.go
@@ -1,0 +1,7 @@
+package repositories
+
+import "gorm.io/gorm"
+
+type DBRepository interface {
+	Conn() *gorm.DB
+}
