+++ app/src/application/repositories/user_repository.go
@@ -2,10 +2,11 @@ package repositories
 
 import (
 	"github.com/takeuchi-shogo/clean-architecture-golang/src/entities"
+	"gorm.io/gorm"
 )
 
 type UserRepository interface {
-	FindByID(userID int) (user entities.Users, err error)
-	FindByScreenName(screenName string) (user entities.Users, err error)
-	Create(user entities.Users) (createdUser entities.Users, err error)
+	FindByID(db *gorm.DB, userID int) (user entities.Users, err error)
+	FindByScreenName(db *gorm.DB, screenName string) (user entities.Users, err error)
+	Create(db *gorm.DB, user entities.Users) (createdUser entities.Users, err error)
 }
