+++ app/src/application/usecases/interactor/product/user_interactor.go
@@ -10,12 +10,15 @@ import (
 )
 
 type UserInteractor struct {
+	DB            repositories.DBRepository
 	User          repositories.UserRepository
 	UserPresenter output.UserInteractorOutput
 }
 
 func (i *UserInteractor) Get(id int) (user entities.Users, resultStatus *usecases.ResultStatus) {
-	user, err := i.User.FindByID(id)
+	db := i.DB.Conn()
+
+	user, err := i.User.FindByID(db, id)
 	if err != nil {
 		return entities.Users{}, usecases.NewResultStatus(400, []string{"users.get"}, errors.New("e"))
 	}
