+++ app/src/infrastructure/cmd/command.go
@@ -1,4 +1,4 @@
-package infrastructure
+package cmd
 
 import (
 	"github.com/takeuchi-shogo/clean-architecture-golang/lib"
@@ -16,7 +16,7 @@ import (
 // }
 
 // main.go の記述をシンプルにするために
-func NewInfrastructure(lib lib.Library) {
+func RunServer(lib lib.Library) {
 
 	config := config.NewConfig(lib.Env)
