+++ docker/mysql/Dockerfile
@@ -4,4 +4,4 @@ FROM mysql:8.0.31
 
 COPY ./mysql/config/my.conf /etc/mysql/conf.d/my.cnf
 # # 起動時に実行してくれるディレクトリ
-# COPY ./mysql/migrations/0001_init_db.sql /docker-entrypoint-initdb.d/0001_init_db.sql
+COPY ./mysql/migrations /docker-entrypoint-initdb.d
