+++ mysql/sql/20220110_create_users.sql
@@ -1,0 +1,12 @@
+CREATE TABLE IF NOT EXISTS `users` (
+	`id` INT(10) UNSIGNED NOT NULL AUTO_INCREMENT,
+	`display_name` varchar(50) NOT NULL,
+	`screen_name` varchar(50) NOT NULL,
+	`email` varchar(30),
+	`password` varchar(255) NOT NULL, 
+	`created_at` int UNSIGNED NOT NULL,
+	`updated_at` int UNSIGNED NOT NULL,
+	PRIMARY KEY (`id`)
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;
+
+DROP TABLE IF EXISTS `users`;
